Ssh accessproxy update (#221)
* fix: [server] support for TRASA PAM in upstream server when accessing from web

* fix: [server] added kb-interactive auth method before password auth.

* feat: [server] added ability to choose service from service name. closes #153

* feat: [server] added test. fixed inconsistency in ssh cli test.  closes #182

* ref: [server] comments added, replaced redundant code with functions
