+++ tests/server/mockdata.sql
@@ -7,7 +7,7 @@ INSERT INTO users (org_id, id, username, first_name, middle_name, last_name, ema
 INSERT INTO services (id, org_id, name, secret_key, passthru, hostname, type, managed_accounts, adhoc, remoteapp_name, session_record, rdp_protocol, native_log, proxy_config, external_provider_name, external_id, external_security_group, distro_name, distro_version, ip_details, created_at, updated_at, deleted_at) VALUES
 	('08d97469-4a2f-46d3-86bc-3005b4c99c6c', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', 'gcp win', '2a094b1fba624b26eaa02f5e2b9f5755ea', false, 'rdp-test-server', 'rdp', '', true, '', false, 'tls', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": true, "upstreamServer": ""}', '', '', '{}', '', '', '{"IPAddr": "0.0.0.0", "defaultGateway": "0.0.0.0", "netMask": ""}', 1596437360, 1596530272, 0),
 	('18d97469-4a2f-46d3-86bc-3005b4c99c6c', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', 'radius test', '3a094b1fba624b26eaa02f5e2b9f5755ea', false, '172.16.238.22', 'radius', '', true, '', false, 'nla', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": true, "upstreamServer": ""}', '', '', '{}', '', '', '{"IPAddr": "0.0.0.0", "defaultGateway": "0.0.0.0", "netMask": ""}', 1596437360, 1596530272, 0),
-	('2fef188a-cc13-438b-8564-2803a072f650', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', '127.0.0.1:2222', 'd9ef5359f13f6f6f6c89b4a9be9958ed13', false, '127.0.0.1:2222', 'ssh', 'root,admin', false, '', true, 'nla', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": true, "upstreamServer": ""}', '', '', '{}', '', '', '{"IPAddr": "0.0.0.0", "defaultGateway": "0.0.0.0", "netMask": ""}', 1596422991, 1596422991, 0),
+	('2fef188a-cc13-438b-8564-2803a072f650', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', 'test-service', 'd9ef5359f13f6f6f6c89b4a9be9958ed13', false, '127.0.0.1:2222', 'ssh', 'root,admin', false, '', true, 'nla', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": true, "upstreamServer": ""}', '', '', '{}', '', '', '{"IPAddr": "0.0.0.0", "defaultGateway": "0.0.0.0", "netMask": ""}', 1596422991, 1596422991, 0),
 	('4ea851b8-6299-4c61-8137-58771aaa8899', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', 'httptest', 'b19f3408628eacee877131a11b96bd71d6', false, 'localhost:3339', 'http', 'admin', false, '', true, 'nla', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": false, "upstreamServer": "http://localhost:3338"}', '', '', '{}', '', '', '{"IPAddr": "", "defaultGateway": "", "netMask": ""}', 1596342013, 1596371470, 0),
 	('7ea551b8-6299-4c61-8137-58771aaa8890', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', 'httptest2', 'f19f3408678eacee877131a11b96bd71d6', false, 'test00.trasa.io', 'http', 'admin', false, '', true, 'nla', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": false, "upstreamServer": "https://gitlab.trasa.io"}', '', '', '{}', '', '', '{"IPAddr": "", "defaultGateway": "", "netMask": ""}', 1596342013, 1596371470, 0),
 	('3ea851b8-6296-4c61-8157-58771aaa8891', '153f7582-5ae2-46ba-8c1c-79ef73fe296e', 'sshtest', 'a19f3408628eacee877131a11b96bd71d7', false, '172.16.238.13', 'ssh', 'admin', false, '', true, 'nla', false, '{"passHostHeader": false, "routeRule": "", "strictTLSValidation": false, "upstreamServer": "https://gitlab.trasa.io"}', '', '', '{}', '', '', '{"IPAddr": "", "defaultGateway": "", "netMask": ""}', 1596342013, 1596371470, 0);
